allocation-api/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-RESTX==1.3.0
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0

# Database
SQLAlchemy==1.4.49
snowflake-connector-python==3.1.1
snowflake-sqlalchemy==1.5.0

# API & Validation
marshmallow==3.20.1
python-dotenv==1.0.0

# HTTP Client for Aladdin
httpx==0.25.0
tenacity==8.2.3

# Async task processing
celery==5.3.1
redis==5.0.0

# Data processing
numpy==1.25.2
scipy==1.11.2
pandas==2.1.0

# Security
cryptography==41.0.4
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1
pytest-cov==4.1.0
faker==19.6.1

# Development
black==23.9.1
flake8==6.1.0
mypy==1.5.1
pre-commit==3.4.0